The WorkSheet component is available
Il componente WorkSheet visualizza la griglia con il tempo speso per progetto per giorno.
Il componente AppComponent è ora il container dell'applicazione ed invoca i componenti necessari a creare il layout della pagina.
