--- src/app/app.component.css
@@ -1,34 +1,0 @@
-td.even {
-    background-color: #f3f3f3;
-    padding: 2px;
-    vertical-align: middle;
-  }
-.odd {
-    background-color: lightgray;
-    padding: 2px;
-    vertical-align: middle;
-  }
-th {
-    background-color: darkgray;
-}
-
-table {
-  table-layout: fixed;
-  border-collapse: collapse;
-  width: 98%; 
-  margin-left: auto;
-  margin-right: auto;
-}
-td {
-  border: 1px solid #FFFFFF;
-  width: 150px;
-}
-td+td.even {
-    width: auto;
-  }
-td+td.odd {
-    width: auto;
-  }
-  body{
-      align-content: center;
-  }
