+++ src/app/app.component.html
@@ -1,19 +1,1 @@
-<h1>Rapportini</h1>
-<table class="table table-striped table-bordered">
-  <tr>
-    <th style="width: 100px;">Progetto</th>
-    <th *ngFor="let d of giorni_del_mese">
-      <p>{{ d }} </p>
-    </th>
-  </tr>
-  <tr *ngFor="let r of righe;let even = even; let odd = odd" [ngClass]="{ odd: odd, even: even }">
-    <td  scope="row"  style="width: 100px;" [ngClass]="{ odd: odd, even: even }">
-      <p>{{ r.progetto.codice }}</p><p>{{ r.progetto.descrizione }}</p>
-    </td>
-    <td *ngFor="let d of r.orari" [ngClass]="{ odd: odd, even: even }">
-        <span  contenteditable="true" (blur)="updateOrario(r.progetto.codice, d, $event)">           
-          {{ d.orario }}
-        </span>
-    </td>
-  </tr>
-</table>
+<app-work-sheet></app-work-sheet>
