+++ src/app/app.module.ts
@@ -3,10 +3,12 @@ import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http'
 import { AppComponent } from './app.component';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+import { WorkSheetComponent } from './work-sheet/work-sheet.component';
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    WorkSheetComponent
   ],
   imports: [
     BrowserModule,
